requirements.txt
# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Database & Cache
sqlalchemy>=2.0.23
redis>=5.0.1
databases[postgresql]>=0.8.0

# Machine Learning Core
pandas>=2.1.4
numpy>=1.26.2
scikit-learn>=1.3.2
joblib>=1.3.2

# ML Enhancement Libraries
xgboost>=2.0.3
lightgbm>=4.1.0
catboost>=1.2.2
optuna>=3.5.0
imbalanced-learn>=0.11.0

# Deep Learning & GNNs
torch>=2.1.2
torch-geometric>=2.4.0

# Quantum ML
pennylane>=0.33.1

# Visualization & Monitoring
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.17.0
streamlit>=1.28.2

# Explainable AI
shap>=0.44.1

# API & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
